Issue ID is a long; updating test to reflect.

+++ Octokit.Tests.Integration/Clients/ProjectCardsClientTests.cs
@@ -453,7 +453,7 @@ public class ProjectCardsClientTests
         return result;
     }
 
-    private static async Task<ProjectCard> CreateIssueCardHelper(IGitHubClient githubClient, int issueId, int columnId)
+    private static async Task<ProjectCard> CreateIssueCardHelper(IGitHubClient githubClient, long issueId, int columnId)
     {
         var newCard = new NewProjectCard(issueId, ProjectCardContentType.Issue);
         var result = await githubClient.Repository.Project.Card.Create(columnId, newCard);
